Add perftsuite command

+++ src/littlewing/game.rs
@@ -127,7 +127,7 @@ impl Game {
         }).fold(String::new(), |r, s| r + s.as_slice()).as_slice()
     }
 
-    pub fn perft(&mut self, depth: uint) -> uint {
+    pub fn perft(&mut self, depth: uint) -> u64 {
         //println!("perft({})", depth);
         if depth == 0 {
             1
@@ -241,25 +241,25 @@ mod tests {
     fn test_perft() {
         let fen = "8/2p5/3p4/KP5r/1R3p1k/8/4P1P1/8 w";
         let mut game = Game::from_fen(fen);
-        assert_eq!(game.perft(1), 14u);
-        //assert_eq!(game.perft(2), 191u);
+        assert_eq!(game.perft(1), 14);
+        //assert_eq!(game.perft(2), 191);
 
         let fen = "r3k2r/Pppp1ppp/1b3nbN/nP6/BBP1P3/q4N2/Pp1P2PP/R2Q1RK1 w";
         let mut game = Game::from_fen(fen);
-        assert_eq!(game.perft(1), 6u);
-        //assert_eq!(game.perft(2), 264u);
+        assert_eq!(game.perft(1), 6);
+        //assert_eq!(game.perft(2), 264);
 
         let fen = "r3k2r/p1ppqpb1/bn2pnp1/3PN3/1p2P3/2N2Q1p/PPPBBPPP/R3K2R w";
         let mut game = Game::from_fen(fen);
-        //assert_eq!(game.perft(1), 48u);
-        //assert_eq!(game.perft(2), 2039u);
-        //assert_eq!(game.perft(3), 97862u);
+        //assert_eq!(game.perft(1), 48);
+        //assert_eq!(game.perft(2), 2039);
+        //assert_eq!(game.perft(3), 97862);
 
         let fen = "rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR w";
         let mut game = Game::from_fen(fen);
-        assert_eq!(game.perft(1), 20u);
-        assert_eq!(game.perft(2), 400u);
-        //assert_eq!(game.perft(3), 8902u);
+        assert_eq!(game.perft(1), 20);
+        assert_eq!(game.perft(2), 400);
+        //assert_eq!(game.perft(3), 8902);
     }
 
     #[test]
